if os.path.exists unlink
git-svn-id: http://web2py.googlecode.com/svn/trunk@581 8fb14a07-5544-0410-8dbc-e3a1454e9f17

+++ gluon/sql.py
@@ -311,7 +311,7 @@ def delete_uploaded_files(table,records,fields=True):
                record.get(fieldname,None):
                 oldname=os.path.join(table._db._folder,
                         '../uploads/',record[fieldname])
-                os.unlink(oldname)
+                if os.path.exists(oldname): os.unlink(oldname)
 
 def cleanup(text):
     if re.compile('[^0-9a-zA-Z_]').findall(text):
